requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.6
idna==2.10
joblib==1.0.0
nltk==3.5
psycopg2==2.8.6
python-rake==1.5.0
pytz==2021.1
rake-nltk==1.0.4
regex==2020.11.13
requests==2.25.1
soupsieve==2.1
sqlparse==0.4.1
tqdm==4.56.0
urllib3==1.26.3
wikipedia==1.4.0
Wikipedia-API==0.5.4
